requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.4
pandas==1.5.1
psycopg2-binary==2.9.5
pyairtable==1.3.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.43
urllib3==1.26.12
matplotlib==3.7.0
oauth2client==4.1.3
pyOpenSSL==23.2.0
gspread==5.10.0
google==3.0.0
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
pandas-gbq==0.23.1
kafka-python==2.0.2